refactor(lib)!: Make rebar3's doCheck opt-in
It's not generally necessary to re-run rebar3's full test suite on a
given OTP version and this is the lion's share of CI time and delays
instantiating local devShells on any cache miss.

At the time of writing the rebar3 project tests with the last 3 OTP
major release:

https://github.com/erlang/rebar3/blob/59238ca1bb4e949f367dac3505c335fb097f208e/.github/workflows/main.yml#L15-L16

BREAKING CHANGE: Users on an older or otherwise "exotic" OTP version <=
24 may experience compatibility issues with `rebar3` (which were already
present) but without the rebar3 test suite informing you during devShell
instantiation.